// src/types/interfaces.ts
export type TextAlignType = 'left' | 'center' | 'right';

export type EmailTypes =
  | '선택 안함'
  | 'naver.com'
  | 'gmail.com'
  | 'daum.net'
  | 'nate.com'
  | 'hanmail.com';

export type FontFamilyTypes =
  | 'extraBold'
  | 'extraLight'
  | 'semiBold'
  | 'regular'
  | 'medium'
  | 'light'
  | 'black'
  | 'bold'
  | 'thin';

export type TextDecorationType =
  | 'dashed'
  | 'double'
  | 'dotted'
  | 'solid'
  | 'line-through'
  | 'overline'
  | 'underline'
  | 'none';

// 월 표기 약자
export type MonthlyAbbreviationType =
  | 'jan'
  | 'feb'
  | 'mar'
  | 'apr'
  | 'may'
  | 'jun'
  | 'jul'
  | 'aug'
  | 'sept'
  | 'oct'
  | 'nov'
  | 'dec'
  | 'jan';